requirements.txt
asyncio==3.4.3 
asyncpg==0.27.0 
cloud-sql-python-connector["asyncpg"]==1.2.3
numpy==1.22.4 
pandas==1.5.3
pgvector==0.1.8
langchain==0.0.196 
transformers==4.30.1
google-cloud-aiplatform==1.26.0